Refactor common Delta Lake test code
